Extract common features of Azure service proxies
The service classes should represent their unique features, and the
common code required to work with them should be extracted to a shared
module. This change does so.

Next steps would be:
 * Update `Graph#get()` to use the cache
 * Move `Graph#get()` to `Service`, replacing the one extracted from
 `Management`
 * Maybe find some smoother way to define the particulars of each class,
 rather than instance variables in `initialize`
