+++ libraries/support/azure.rb
@@ -3,5 +3,6 @@
 require 'support/azure/authentication'
 require 'support/azure/rest'
 require 'support/azure/credentials'
+require 'support/azure/service'
 require 'support/azure/management'
 require 'support/azure/graph'
