+++ libraries/support/azure/management.rb
@@ -1,38 +1,15 @@
 # frozen_string_literal: true
 
-require 'json'
 require 'singleton'
 
 module Azure
   class Management
     include Singleton
-
-    class Cache
-      def initialize
-        @store = {}
-      end
-
-      def fetch(key)
-        @store.fetch(key) { @store[key] = yield }
-      end
-    end
-
-    attr_reader :cache
+    include Service
 
     def initialize
-      @cache = Cache.new
-    end
-
-    def with_cache(cache)
-      @cache = cache
-    end
-
-    def with_client(rest_client, override: false)
-      set_reader(:rest_client, rest_client, override)
-    end
-
-    def for_subscription(subscription_id, override: false)
-      set_reader(:subscription_id, subscription_id, override)
+      @required_attrs = %i(rest_client subscription_id)
+      @page_link_name = 'nextLink'
     end
 
     def activity_log_alert(resource_group, id)
@@ -152,31 +129,6 @@ module Azure
       "#{"/resourceGroups/#{resource_group}" if resource_group}" \
       "#{'/providers' if provider}" \
       "/#{location}/"
-    end
-
-    def get(url:, api_version:)
-      confirm_configured!
-
-      cache.fetch(url) do
-        body = rest_client.get(url, params: { 'api-version' => api_version }).body
-        body.fetch('value', body)
-      end
-    end
-
-    def confirm_configured!
-      %i(rest_client subscription_id).each do |name|
-        next if respond_to?(name)
-
-        raise "Set #{name} before making requests"
-      end
-    end
-
-    def set_reader(name, value, override)
-      return self if respond_to?(name) && !override
-
-      define_singleton_method(name) { value }
-
-      self
     end
   end
 end
